feat(is_equivalent): Allow notes from other tunings to be compared
Implementing is_equivalent over pc_index would need an additional
check if tunings are the same, however that would restrict us
without necessity. Rather what we actual want to know is if the
frequencies are the same when the objects are normalized to the
zero base interval.

The new implementation allows is_equivalent to not only compare
notes and pitches, but also notes from different tunings, e.g.
to check if pitch index 3 from 12-EDO and pitch index 30 from
24-EDO are equivalent