test(frequencies): Stricter object comparison
The == test might give false impressions on whether the type of
the arithmetic result is correct. This happened for example when
sympy simply converted Frequency and FrequencyRatios heuristically
into its own types resulting in arithmetic results that did not
have the expected xenharmlib type but the sympy type.

The additional tests make sure that the resulting type is correct.