chore: Removed Google site verification token
File was not served on RTD anyways and everything points to Google
just being very very slow to index RTD and PyPI correctly. There
seems to be no misconfiguration on my part, since - and I cannot
believe I say this, someone pinch me this must be a dream - the
site was correctly indexed by Bing.