refactor: don't use add_note / add_pitch on scales
because scales need to preserve order we need to insort every
new element that is added, however since cpython does not use
linked lists but arrays, the O(log n) of insorting is sadly
overshadowed by the list copy operation which is O(n).
this means it is way more costly to insort into a scale
than to sort once on construction. this commit replaces
all code using add_note/add_pitch to build up scales
with simple list building and subsequent(!) scale
construction.

in the future add_note / add_pitch should be removed anyway
because it does not follow the overall functional paradigm
of the library.