refactor: Introduced abstract interval class
Introduced abstract interval classes Interval and SDInterval
that define a common interface for both note intervals and
pitch intervals.

Deprecated tuning.pitch_interval in favor of .interval
Deprecated notation.note_interval in favor of .interval
Deprecated note_interval.from_notes in favor of .from_source_and_target
Deprecated pitch_interval.from_pitches in favor of .from_source_and_target

BREAKING CHANGE: Changed the constructor parameters of intervals to
align them with the newly introduced base classes. Subclassing
interval types are not yet part of the documentation, so this
should not ruffle too many feathers even if this is a
minor-version release. It should still be mentioned in
the changelog however.