fix: temporarily removed IntervalSeq type annotation
The import lead to a circular dependency of the scale and
interval_seq module, similar to the problem of circular
dependency between pitches and intervals. In this commit
the type annotation has been removed, which is not ideal,
however putting interval sequences and scales together
in one module like with pitch and interval is also not
a good organization of code. there needs to be a better
solution in the future.