fix: PeriodicScale.is_equivalent conceptual error
is_equivalent was documented in a way that it would allow for two
similar but distinct interpretations. while one part of the library
assumed interpretation A (is_equivalent tests for pitch class set
equality), another assumed interpretation B (is_equivalent tests
for base interval index shift). While A does not care about order
or length equality B does. In many cases A and B produce the same
results, that's why the error was found so late.

this is a bad situation, because it is a very philosophical
question whether this is a bug at all. the implementation
of is_equivalent is hereby changed from interpretation A to
interpretation B, because the name is_equivalent implies
somewhat more strongly that two scales with unequal length
should not be considered equivalent.

for interpretation A a new function (is_pcs_equivalent)
was introduced and all calls that assumed interpretation
A were changed. in addition the docstrings were adjusted
so the difference between the two functions is clear.