fix: is_equivalency fiasko
after careful consideration the is_equivalent method for scales
should NOT change behavior, rather it should be marked as
deprecated and replaced by two distinct methods

since the scale object is both set and sequence there is an
argument to be made both for the sequence equivalence and
the set equivalence behavior.

is_equivalent is now deprecated and is_set_equivalent
and is_seq_equivalent is introduced. is_equivalent
was changed back to old 0.2.x behavior